pyproject.toml
[project]
name = "kombyphantike"
version = "0.1.0"
description = "Diachronic analysis of Greek vocabulary for the antifluent philologist"
authors = [
    {name = "I. Mert Inan",email = "124540268+Alleksenes@users.noreply.github.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"  # Force a version that supports Torch/SpaCy
spacy = "^3.7"
stanza = "^1.8"
pandas = "^2.2"
requests = "^2.31"
openpyxl = "^3.1" # For reading Excel/Google Sheets exports
numpy = "<2.0.0"
sentence-transformers = "^5.2.0"
gspread = "^6.2.1"
oauth2client = "^4.1.3"
gtts = "^2.5.4"
pydub = "^0.25.1"
python-dotenv = "^1.2.1"
grc-odycy-joint-sm = {path = "./grc_odycy_joint_sm-0.1.0-py3-none-any.whl"}
streamlit = "^1.52.2"
google-generativeai = "^0.8.6"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29" # So you can use Jupyter notebooks in VS Code

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"